+++ src/container/header/mod.rs
@@ -4,27 +4,27 @@ mod testbench;
 use crate::ast::line::Line;
 use crate::utils::{drop_while, lowercase, take_and_drop_while};
 
-pub fn parse_header(line: &Line) -> (usize, Vec<u16>) {  // (level, content)
-    let (sharps, sharps_removed) = take_and_drop_while(&line.content, '#' as u16);
-    let indents_removed = drop_while(&sharps_removed, ' ' as u16);
+pub fn parse_header(line: &Line) -> (usize, Vec<u32>) {  // (level, content)
+    let (sharps, sharps_removed) = take_and_drop_while(&line.content, '#' as u32);
+    let indents_removed = drop_while(&sharps_removed, ' ' as u32);
 
     (sharps.len(), indents_removed)
 }
 
-pub fn normalize_header(content: &[u16]) -> Vec<u16> {
+pub fn normalize_header(content: &[u32]) -> Vec<u32> {
     content.iter().filter(
-        |c| '0' as u16 <= **c && **c <= '9' as u16
-        || 'a' as u16 <= **c && **c <= 'z' as u16
-        || 'A' as u16 <= **c && **c <= 'Z' as u16
-        || '가' as u16 <= **c && **c <= '힣' as u16
-        || 'ㄱ' as u16 <= **c && **c <= 'ㅣ' as u16  // Korean
-        || 'ぁ' as u16 <= **c && **c <= 'ヺ' as u16  // Japanese
-        || '-' as u16 == **c || '_' as u16 == **c || ' ' as u16 == **c
+        |c| '0' as u32 <= **c && **c <= '9' as u32
+        || 'a' as u32 <= **c && **c <= 'z' as u32
+        || 'A' as u32 <= **c && **c <= 'Z' as u32
+        || '가' as u32 <= **c && **c <= '힣' as u32
+        || 'ㄱ' as u32 <= **c && **c <= 'ㅣ' as u32  // Korean
+        || 'ぁ' as u32 <= **c && **c <= 'ヺ' as u32  // Japanese
+        || '-' as u32 == **c || '_' as u32 == **c || ' ' as u32 == **c
     ).map(
-        |c| if *c == ' ' as u16 {
-            '-' as u16
+        |c| if *c == ' ' as u32 {
+            '-' as u32
         } else {
             lowercase(c)
         }
-    ).collect::<Vec<u16>>()
+    ).collect::<Vec<u32>>()
 }
