+++ src/container/list/tasklist.rs
@@ -3,28 +3,28 @@ use crate::utils::drop_while;
 
 // line.content doesn't contain bullet/marker
 pub fn parse_task_list(line: &Line) -> (Line, Option<TaskMarker>) {
-    let whitespace_trimed = drop_while(&line.content, ' ' as u16);
+    let whitespace_trimed = drop_while(&line.content, ' ' as u32);
 
     let is_task_list = whitespace_trimed.len() > 4
-        && whitespace_trimed[0] == '[' as u16
-        && whitespace_trimed[2] == ']' as u16
-        && whitespace_trimed[3] == ' ' as u16
+        && whitespace_trimed[0] == '[' as u32
+        && whitespace_trimed[2] == ']' as u32
+        && whitespace_trimed[3] == ' ' as u32
         && (
-            whitespace_trimed[1] == ' ' as u16
-            || whitespace_trimed[1] == 'x' as u16
-            || whitespace_trimed[1] == 'X' as u16
-            || whitespace_trimed[1] == '^' as u16
+            whitespace_trimed[1] == ' ' as u32
+            || whitespace_trimed[1] == 'x' as u32
+            || whitespace_trimed[1] == 'X' as u32
+            || whitespace_trimed[1] == '^' as u32
     );
 
     if !is_task_list {
         (line.clone(), None)
     }
 
-    else if whitespace_trimed[1] == ' ' as u16 {
+    else if whitespace_trimed[1] == ' ' as u32 {
         (Line::new(whitespace_trimed[4..].to_vec(), line.indent), Some(TaskMarker::Unchecked))
     }
 
-    else if whitespace_trimed[1] == '^' as u16 {
+    else if whitespace_trimed[1] == '^' as u32 {
         (Line::new(whitespace_trimed[4..].to_vec(), line.indent), Some(TaskMarker::Triangle))
     }
