+++ src/inline/footnote/predicate.rs
@@ -3,13 +3,13 @@ use crate::inline::link::normalize_link_label;
 use crate::utils::get_bracket_end_index;
 use std::collections::HashMap;
 
-pub fn is_valid_footnote_label(content: &[u16]) -> bool {
-    content[0] == '^' as u16 && content[1..content.len()].iter().all(is_valid_footnote_character)
+pub fn is_valid_footnote_label(content: &[u32]) -> bool {
+    content[0] == '^' as u32 && content[1..content.len()].iter().all(is_valid_footnote_character)
 }
 
-pub fn read_footnote(content: &[u16], index: usize, footnote_references: &HashMap<Vec<u16>, Footnote>) -> Option<usize> {  // footnote_index
+pub fn read_footnote(content: &[u32], index: usize, footnote_references: &HashMap<Vec<u32>, Footnote>) -> Option<usize> {  // footnote_index
 
-    if content[index] == '[' as u16 {
+    if content[index] == '[' as u32 {
 
         match get_bracket_end_index(content, index) {
             Some(bracket_end_index) => {
@@ -32,12 +32,12 @@ pub fn read_footnote(content: &[u16], index: usize, footnote_references: &HashMa
 
 }
 
-fn is_valid_footnote_character(c: &u16) -> bool {
+fn is_valid_footnote_character(c: &u32) -> bool {
 
-    '0' as u16 <= *c && *c <= '9' as u16
-    || 'a' as u16 <= *c && *c <= 'z' as u16
-    || 'A' as u16 <= *c && *c <= 'Z' as u16
-    || '가' as u16 <= *c && *c <= '힣' as u16  // Korean
-    || 'ㄱ' as u16 <= *c && *c <= 'ㅣ' as u16  // Korean
-    || 'ぁ' as u16 <= *c && *c <= 'ヺ' as u16  // Japanese
+    '0' as u32 <= *c && *c <= '9' as u32
+    || 'a' as u32 <= *c && *c <= 'z' as u32
+    || 'A' as u32 <= *c && *c <= 'Z' as u32
+    || '가' as u32 <= *c && *c <= '힣' as u32  // Korean
+    || 'ㄱ' as u32 <= *c && *c <= 'ㅣ' as u32  // Korean
+    || 'ぁ' as u32 <= *c && *c <= 'ヺ' as u32  // Japanese
 }
