+++ src/inline/link/mod.rs
@@ -8,6 +8,6 @@ mod testbench;
 use crate::utils::{collapse_whitespaces, lowercase, strip_whitespaces};
 
 // [  F  OO ] -> [f oo]
-pub fn normalize_link_label(original: &[u16]) -> Vec<u16> {
+pub fn normalize_link_label(original: &[u32]) -> Vec<u32> {
     strip_whitespaces(&collapse_whitespaces(original)).iter().map(lowercase).collect()
 }
