+++ src/inline/link/predicate.rs
@@ -10,16 +10,16 @@ use std::collections::HashMap;
 
 // [foo](address)
 pub fn read_direct_link(
-    content: &[u16],
+    content: &[u32],
     index: usize,
-    link_references: &HashMap<Vec<u16>, Vec<u16>>
-) -> Option<(Vec<u16>, Vec<u16>, usize)> {  // Option<(link_text, link_destination, last_index)>
+    link_references: &HashMap<Vec<u32>, Vec<u32>>
+) -> Option<(Vec<u32>, Vec<u32>, usize)> {  // Option<(link_text, link_destination, last_index)>
 
-    if content[index] == '[' as u16 {
+    if content[index] == '[' as u32 {
 
         match get_bracket_end_index(content, index) {
             Some(bracket_end_index) => if bracket_end_index + 1 >= content.len()
-                || content[bracket_end_index + 1] != '(' as u16
+                || content[bracket_end_index + 1] != '(' as u32
             {
                 None
             } else {
@@ -55,13 +55,13 @@ pub fn read_direct_link(
 
 // [foo][bar]
 // [foo][]
-pub fn read_reference_link(content: &[u16], index: usize, link_references: &HashMap<Vec<u16>, Vec<u16>>) -> Option<(Vec<u16>, Vec<u16>, usize)> {  // Option<(link_text, link_label, last_index)>
+pub fn read_reference_link(content: &[u32], index: usize, link_references: &HashMap<Vec<u32>, Vec<u32>>) -> Option<(Vec<u32>, Vec<u32>, usize)> {  // Option<(link_text, link_label, last_index)>
 
-    if content[index] == '[' as u16 {
+    if content[index] == '[' as u32 {
 
         match get_bracket_end_index(content, index) {
             Some(bracket_end_index) => if bracket_end_index + 1 >= content.len()
-                || content[bracket_end_index + 1] != '[' as u16
+                || content[bracket_end_index + 1] != '[' as u32
             {
                 None
             } else {
@@ -71,7 +71,7 @@ pub fn read_reference_link(content: &[u16], index: usize, link_references: &Hash
                         let link_text = &content[index + 1..bracket_end_index];  // `foo` in `[foo][bar]`
                         let mut link_label = &content[bracket_end_index + 2..second_bracket_end_index];  // `bar` in `[foo][bar]`
 
-                        if link_label.len() > 2 && link_label[0] == '[' as u16 {
+                        if link_label.len() > 2 && link_label[0] == '[' as u32 {
                             // `[link][[br]]` is a shortcut reference link followed by a macro
                             match get_bracket_end_index(&link_label, 0) {
                                 Some(i) if i == link_label.len() - 1 => {
@@ -113,22 +113,22 @@ pub fn read_reference_link(content: &[u16], index: usize, link_references: &Hash
 // [foo]
 // in case `foo` is a valid reference and `bar` is not, `[foo][bar]` is an invalid reference.
 // if parenthesises or brackets follow `[foo]`, it's not regarded as a shortcut reference whether or not the references are valid.
-pub fn read_shortcut_reference_link(content: &[u16], index: usize, link_references: &HashMap<Vec<u16>, Vec<u16>>) -> Option<(Vec<u16>, usize)> {  // Option<(link_text, last_index)>
+pub fn read_shortcut_reference_link(content: &[u32], index: usize, link_references: &HashMap<Vec<u32>, Vec<u32>>) -> Option<(Vec<u32>, usize)> {  // Option<(link_text, last_index)>
 
-    if content[index] == '[' as u16 {
+    if content[index] == '[' as u32 {
 
         match get_bracket_end_index(content, index) {
             Some(bracket_end_index) => {
 
                 // if a shortcut reference is followed by a balanced [] or (), the reference is invalid
                 if bracket_end_index + 1 == content.len()
-                    || (content[bracket_end_index + 1] != '(' as u16
-                    && content[bracket_end_index + 1] != '[' as u16)
+                    || (content[bracket_end_index + 1] != '(' as u32
+                    && content[bracket_end_index + 1] != '[' as u32)
                 {
                     //
                 }
 
-                else if content[bracket_end_index + 1] == '(' as u16 {
+                else if content[bracket_end_index + 1] == '(' as u32 {
 
                     match get_parenthesis_end_index(content, bracket_end_index + 1) {
                         Some(_) => {return None;}
@@ -137,13 +137,13 @@ pub fn read_shortcut_reference_link(content: &[u16], index: usize, link_referenc
 
                 }
 
-                else {  // content[bracket_end_index + 1] == '[' as u16
+                else {  // content[bracket_end_index + 1] == '[' as u32
 
                     // `[link][[br]]` is a shortcut reference link followed by a macro
                     match get_bracket_end_index(content, bracket_end_index + 1) {
                         Some(second_bracket_end_index) => {
 
-                            if second_bracket_end_index == bracket_end_index + 2 || content[bracket_end_index + 2] != '[' as u16 {
+                            if second_bracket_end_index == bracket_end_index + 2 || content[bracket_end_index + 2] != '[' as u32 {
                                 return None;
                             }
 
@@ -188,39 +188,39 @@ pub fn read_shortcut_reference_link(content: &[u16], index: usize, link_referenc
 
 // [foo]: address
 // it assumes that the content is a valid link_reference_definition
-pub fn read_link_reference(content: &[u16]) -> (Vec<u16>, Vec<u16>) {  // (link_label, link_destination)
+pub fn read_link_reference(content: &[u32]) -> (Vec<u32>, Vec<u32>) {  // (link_label, link_destination)
 
     let bracket_end_index = get_bracket_end_index(content, 0).unwrap();
     let link_label = content[1..bracket_end_index].to_vec();
-    let link_destination = render_html_escapes(&drop_while(&content[bracket_end_index + 2..content.len()], ' ' as u16));
+    let link_destination = render_html_escapes(&drop_while(&content[bracket_end_index + 2..content.len()], ' ' as u32));
 
     (link_label, link_destination)
 }
 
-fn is_valid_link_text(content: &[u16], link_references: &HashMap<Vec<u16>, Vec<u16>>) -> bool {
+fn is_valid_link_text(content: &[u32], link_references: &HashMap<Vec<u32>, Vec<u32>>) -> bool {
 
     !contains_link(content, link_references)  // it makes sure that the links are not nested
     && (content.len() == 0
-    || if content[0] == '[' as u16
-        && content[content.len() - 1] == ']' as u16
+    || if content[0] == '[' as u32
+        && content[content.len() - 1] == ']' as u32
     {
 
         // [...] -> link
         // [[...]] -> macro
         // [[[...]]] -> macro in a link
-        content.len() > 4 && content[1] == '[' as u16 && content[content.len() - 2] == ']' as u16
+        content.len() > 4 && content[1] == '[' as u32 && content[content.len() - 2] == ']' as u32
     } else {
         true
     })
 }
 
-pub fn is_valid_link_label(content: &[u16]) -> bool {
+pub fn is_valid_link_label(content: &[u32]) -> bool {
 
     // [...] -> link
     // [[...]] -> macro
     // [[[...]]] -> macro in a link
-    if content.len() > 1 && content[0] == '[' as u16 && content[content.len() - 1] == ']' as u16 {
-        content.len() > 4 && content[1] == '[' as u16 && content[content.len() - 2] == ']' as u16
+    if content.len() > 1 && content[0] == '[' as u32 && content[content.len() - 1] == ']' as u32 {
+        content.len() > 4 && content[1] == '[' as u32 && content[content.len() - 2] == ']' as u32
     }
 
     else {
@@ -229,17 +229,17 @@ pub fn is_valid_link_label(content: &[u16]) -> bool {
 
 }
 
-pub fn is_valid_link_destination(content: &[u16]) -> bool {
+pub fn is_valid_link_destination(content: &[u32]) -> bool {
     content.iter().all(is_valid_url_character)
 }
 
-fn contains_link(content: &[u16], link_references: &HashMap<Vec<u16>, Vec<u16>>) -> bool {
+fn contains_link(content: &[u32], link_references: &HashMap<Vec<u32>, Vec<u32>>) -> bool {
 
-    match content.iter().position(|c| *c == '[' as u16) {
+    match content.iter().position(|c| *c == '[' as u32) {
         // this function is used to remove nested links
         // so images are fine
         Some(index)
-            if index == 0 || content[index - 1] != '!' as u16 =>
+            if index == 0 || content[index - 1] != '!' as u32 =>
         {
             read_direct_link(content, index, link_references).is_some()
             || read_reference_link(content, index, link_references).is_some()
@@ -251,16 +251,16 @@ fn contains_link(content: &[u16], link_references: &HashMap<Vec<u16>, Vec<u16>>)
 
 }
 
-fn is_valid_url_character(character: &u16) -> bool {
-    '-' as u16 <= *character && *character <= ';' as u16
-    || 'a' as u16 <= *character && *character <= 'z' as u16
-    || '?' as u16 <= *character && *character <= 'Z' as u16
-    || '가' as u16 <= *character && *character <= '힣' as u16  // korean
-    || 'ㄱ' as u16 <= *character && *character <= 'ㅣ' as u16  // korean
-    || 'ぁ' as u16 <= *character && *character <= 'ヺ' as u16  // japanese
-    || '!' as u16 == *character || '=' as u16 == *character
-    || '+' as u16 == *character || '&' as u16 == *character
-    || '%' as u16 == *character || '_' as u16 == *character
-    || '#' as u16 == *character || '$' as u16 == *character
-    || '+' as u16 == *character
+fn is_valid_url_character(character: &u32) -> bool {
+    '-' as u32 <= *character && *character <= ';' as u32
+    || 'a' as u32 <= *character && *character <= 'z' as u32
+    || '?' as u32 <= *character && *character <= 'Z' as u32
+    || '가' as u32 <= *character && *character <= '힣' as u32  // korean
+    || 'ㄱ' as u32 <= *character && *character <= 'ㅣ' as u32  // korean
+    || 'ぁ' as u32 <= *character && *character <= 'ヺ' as u32  // japanese
+    || '!' as u32 == *character || '=' as u32 == *character
+    || '+' as u32 == *character || '&' as u32 == *character
+    || '%' as u32 == *character || '_' as u32 == *character
+    || '#' as u32 == *character || '$' as u32 == *character
+    || '+' as u32 == *character
 }
