+++ src/inline/math/validate.rs
@@ -5,16 +5,16 @@ use std::collections::HashSet;
 lazy_static! {
 
     // it's only used by `is_valid`
-    static ref FUNCTION_NAMES: HashSet<Vec<u16>> = {
+    static ref FUNCTION_NAMES: HashSet<Vec<u32>> = {
         ZERO_ARG_FUNCTIONS
-            .union(&ONE_ARG_FUNCTIONS).map(|f| f.clone()).collect::<HashSet<Vec<u16>>>()
-            .union(&TWO_ARG_FUNCTIONS).map(|f| f.clone()).collect::<HashSet<Vec<u16>>>()
-            .union(&THREE_ARG_FUNCTIONS).map(|f| f.clone()).collect::<HashSet<Vec<u16>>>()
-            .union(&FIVE_ARG_FUNCTIONS).map(|f| f.clone()).collect::<HashSet<Vec<u16>>>()
+            .union(&ONE_ARG_FUNCTIONS).map(|f| f.clone()).collect::<HashSet<Vec<u32>>>()
+            .union(&TWO_ARG_FUNCTIONS).map(|f| f.clone()).collect::<HashSet<Vec<u32>>>()
+            .union(&THREE_ARG_FUNCTIONS).map(|f| f.clone()).collect::<HashSet<Vec<u32>>>()
+            .union(&FIVE_ARG_FUNCTIONS).map(|f| f.clone()).collect::<HashSet<Vec<u32>>>()
     };
 }
 
-pub fn is_valid(word: &[u16], arguments: &Vec<Vec<u16>>) -> bool {
+pub fn is_valid(word: &[u32], arguments: &Vec<Vec<u32>>) -> bool {
 
     is_space(word)
     || FUNCTION_NAMES.contains(word) && (
